Game Release
Release of the game created in the practical project. I started this project way later than it should have been, and that is why it's not as good as it could be.
Main point and mechanic of the game is implemented, the more you click, the greater will be your speed value.
Firebase updates the value it has stored in it correctly, at every update.
Even tho this project is very simple, I had a good time developing it and it helped me improve on my skills and knownledge using Android Studio.